# Change behavior of AttributeValue oneof value field accessors