Make it consistent with Metric oneof data field: raise panic in case of type mismatch instead of returning a zero value.

The following methods are deprecated in favor of the new accessors:
- StringVal -> StringValue
- IntVal -> IntValue
- DoubleVal -> DoubleValue
- BoolVal -> BoolValue
- MapVal -> MapValue
- SliceVal -> SliceValue
- BytesVal -> BytesValue

Corresponding setter methods are also renamed for consistency without changing its logic:
- SetStringVal -> SetStringValue
- SetIntVal -> SetIntValue
- SetDoubleVal -> SetDoubleValue
- SetBoolVal -> SetBoolValue
- SetBytesVal -> SetBytesValue

Updates: https://github.com/open-telemetry/opentelemetry-collector/issues/4775

## Benchmarks

goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/collector/model/internal/pdata
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz

### Before

```
BenchmarkAttributeValueStringDeprecatedAccessor-16               	1000000000	         0.6580 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueIntDeprecatedAccessor-16                  	1000000000	         0.6568 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueDoubleDeprecatedAccessor-16               	1000000000	         0.6514 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBoolDeprecatedAccessor-16                 	1000000000	         0.7169 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBytesDeprecatedAccessor-16                	1000000000	         0.6539 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueMapDeprecatedAccessor-16                  	487263888	         2.425 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueArrayDeprecatedAccessor-16                	508004030	         2.372 ns/op	       0 B/op	       0 allocs/op
```

### After

```
BenchmarkAttributeValueStringAccessor-16                         	1000000000	         0.6713 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueIntAccessor-16                            	1000000000	         0.6626 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueDoubleAccessor-16                         	1000000000	         0.6548 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBoolAccessor-16                           	1000000000	         0.7929 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBytesAccessor-16                          	1000000000	         0.6972 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueMapAccessor-16                            	1000000000	         0.9166 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueArrayAccessor-16                          	1000000000	         0.9209 ns/op	       0 B/op	       0 allocs/op
```