@bogdandrutu 2022-02-25
@dmitryax if we break this class anyway. Should we do the rename where the funcs have the new behavior, and the deprecated (old object) keeps the old behavior and old name for the funcs?