@dmitryax 2022-02-26
> The idea being the new struct AnyValue (based on the suggestion) should have only the new getters.

Makes sense. I'll combine other renames related to AttributeValue within this change