@dmitryax 2022-03-04
> @dmitryax if we break this class anyway. Should we do the rename where the funcs have the new behavior, and the deprecated (old object) keeps the old behavior and old name for the funcs?

@bogdandrutu after giving it a second thought I would rather keep this PR as is and will change `AttributeValue` to `AnyValue` in the next PR, because mixing method renames with the struct name change makes it more complicated to follow IMO