@Aneurysm9 2022-03-11
Should we return an error instead of `panic()`ing?  I'm not sure it is a good idea to cause an application to crash due to a mismatched expectation that can be handled gracefully.