@dmitryax 2022-03-11
> Should we return an error instead of panic()ing? I'm not sure it is a good idea to cause an application to crash due to a mismatched expectation that can be handled gracefully.

We do allow panic in other places of pdata already. It's always expected that the type is checked before accessing the data. This is how it's being used almost everywhere in contrib (I'm updating couple of places where it's missed).

If we return error, 
1. It'll complicate the clients. As I mentioned, we check the type everywhere before accessing the data, so it'll leave all the clients omitting errors like `val, _ := ...Value()` making the code look unpleasant. Or we would have to add redundant error checks which will be never be executed given the type is validated before the call.
2. It'll bring back the logic that we are replacing in this PR with all the proto generated conditions + returning an additional error. I think it'll make it even less performant than before.