@Aneurysm9 2022-03-11
> We do allow panic in other places of pdata already. It's always expected that the type is checked before accessing the data. This is how it's being used almost everywhere in contrib (I'm updating couple of places where it's missed).

I'm not sure that "we do <undesirable thing> elsewhere" is a solid argument for doing it in new places.  Indeed, I'm not sure I've seen any reason at all offered for why we want to do this potentially very disruptive change.

> If we return error,
>
>    It'll complicate the clients. As I mentioned, we check the type everywhere before accessing the data, so it'll leave all the clients omitting errors like val, _ := ...Value() making the code look unpleasant. Or we would have to add redundant error checks which will be never be executed given the type is validated before the call.

We apparently don't type check everywhere since you've opened at least one issue to add missing type checking.  Indeed, I'm not sure that we type check in most places or even anywhere.  A cursory search for `StringVal()` in the contrib returns dozens of hits where attribute string values are accessed unchecked, or only checked by looking for an empty string after access.  Each of these places would become a landmine waiting to explode a running process with the attendant potential for losing data.

As for whether it will complicate clients, I don't agree.  Clients either have to check the attribute type first or they have  to check for an error or zero value after access.  In any scenario there is checking that needs to happen, but only in one scenario does failing to check every single time result in a dead process and lost data.

>    It'll bring back all the proto generated conditions logic that we are replacing in this PR + an additional error. I think it'll make it even less performant than before.

Do you have benchmarks for this?  This is a significant change that should be made after testing and proving the benefits, not on instinct.