@dmitryax 2022-03-11
> Do you have benchmarks for this? This is a significant change that should be made after testing and proving the benefits, not on instinct.

Makes sense. I will provide benchmarks for the 3 cases so we can discuss the API based on that:
1. current implementation
2. this PR
3. value + error