@dmitryax 2022-03-13
@Aneurysm9 I tried the approach that you suggested here https://github.com/open-telemetry/opentelemetry-collector/pull/4987

The benchmarks are the following:

**Before**

```
BenchmarkAttributeValueStringDeprecatedAccessor-16               	1000000000	         0.6580 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueIntDeprecatedAccessor-16                  	1000000000	         0.6568 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueDoubleDeprecatedAccessor-16               	1000000000	         0.6514 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBoolDeprecatedAccessor-16                 	1000000000	         0.7169 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBytesDeprecatedAccessor-16                	1000000000	         0.6539 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueMapDeprecatedAccessor-16                  	487263888	         2.425 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueArrayDeprecatedAccessor-16                	508004030	         2.372 ns/op	       0 B/op	       0 allocs/op
```

**With panic (This PR)**

```
BenchmarkAttributeValueStringAccessor-16                         	1000000000	         0.6713 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueIntAccessor-16                            	1000000000	         0.6626 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueDoubleAccessor-16                         	1000000000	         0.6548 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBoolAccessor-16                           	1000000000	         0.7929 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueBytesAccessor-16                          	1000000000	         0.6972 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueMapAccessor-16                            	1000000000	         0.9166 ns/op	       0 B/op	       0 allocs/op
BenchmarkAttributeValueArrayAccessor-16                          	1000000000	         0.9209 ns/op	       0 B/op	       0 allocs/op
```

**With error (https://github.com/open-telemetry/opentelemetry-collector/pull/4987)**

```
BenchmarkAttributeValueStringAccessor-16                            353806957            3.464 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueIntAccessor-16                               346630108            3.479 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueDoubleAccessor-16                            356147574            3.358 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueBoolAccessor-16                              350702415            3.472 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueBytesAccessor-16                             331781389            3.677 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueMapAccessor-16                               341309820            3.559 ns/op           0 B/op          0 allocs/op
BenchmarkAttributeValueArrayAccessor-16                             346380688            3.493 ns/op           0 B/op          0 allocs/op
```