@bogdandrutu 2022-03-16
@dmitryax I think this needs to be closed based on the decision in https://github.com/open-telemetry/opentelemetry-collector/issues/4775?